Code voor het draggen in de DragController class verplaatst + paar comments toegevoegd
